Optimize Admin Module by creating separate functions
